# apps/web/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/users/withdraw/finalize": {
      "post": {
        "operationId": "AdminUsersController_finalizeWithdrawals",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "유예 만료 탈퇴 확정 (cron 수동 실행)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "AdminUsersController_findAll",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 80,
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "user",
                "admin"
              ]
            }
          },
          {
            "name": "inactiveDays",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "activeToday",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUsersPageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "사용자 목록 · q·필터·cursor",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/admin/rooms": {
      "get": {
        "operationId": "AdminRoomsController_list",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 40,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "closed",
                "archived"
              ]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRoomsPageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 목록 (운영) · q·status·cursor",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/rooms/messages/{messageId}": {
      "get": {
        "operationId": "AdminRoomsController_getMessage",
        "parameters": [
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 메시지 상세 (신고 원문)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/rooms/{id}/messages/{messageId}": {
      "delete": {
        "operationId": "AdminRoomsController_deleteMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 메시지 운영 삭제 (tombstone)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/rooms/{id}": {
      "patch": {
        "operationId": "AdminRoomsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 상태 · active/closed/archived · 닫기·보관 시 reason",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/dms": {
      "get": {
        "operationId": "DmsController_listMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DmListItemResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "열린 DM 목록",
        "tags": [
          "DMs"
        ]
      },
      "post": {
        "operationId": "DmsController_openOrGet",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenDmDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "DM 열기 또는 요청 (없으면 생성)",
        "tags": [
          "DMs"
        ]
      }
    },
    "/dms/requests": {
      "get": {
        "operationId": "DmsController_listRequests",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DmRequestItemResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "받은 DM 목록",
        "tags": [
          "DMs"
        ]
      }
    },
    "/dms/{id}": {
      "get": {
        "operationId": "DmsController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DmDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "DM 상세 (상대·status)",
        "tags": [
          "DMs"
        ]
      }
    },
    "/dms/{id}/accept": {
      "post": {
        "operationId": "DmsController_accept",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "DM 요청 수락",
        "tags": [
          "DMs"
        ]
      }
    },
    "/dms/{id}/decline": {
      "post": {
        "operationId": "DmsController_decline",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "DM 요청 거절",
        "tags": [
          "DMs"
        ]
      }
    },
    "/dms/{id}/messages": {
      "get": {
        "operationId": "DmsController_listMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DmMessageResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "DM 메시지 목록",
        "tags": [
          "DMs"
        ]
      },
      "post": {
        "operationId": "DmsController_sendMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendDmMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DmMessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "DM 메시지 전송",
        "tags": [
          "DMs"
        ]
      }
    },
    "/dms/{id}/read": {
      "post": {
        "operationId": "DmsController_markRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "DM 메시지 읽음 처리",
        "tags": [
          "DMs"
        ]
      }
    },
...
    "/rooms/{id}": {
      "get": {
        "operationId": "RoomsController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 상세 조회",
        "tags": [
          "Rooms"
        ]
      },
      "patch": {
        "operationId": "RoomsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "방 설정 수정(방장)",
        "tags": [
          "Rooms"
        ]
      }
    },
...
    "/rooms/{id}/messages/{messageId}": {
      "delete": {
        "operationId": "RoomsController_deleteMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "메시지 전체에서 삭제(작성자·방장)",
        "tags": [
          "Rooms"
        ]
      }
    },
...
  },
  "info": {
    "title": "Music Community API",
    "description": "Music Community API description",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "POST /auth/login 또는 /auth/register 후 발급받은 토큰을 사용하세요."
      }
    },
    "schemas": {
...
      "AuthorSnippetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nickname": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "nickname"
        ]
      },
...
      "AdminUserCountsDto": {
        "type": "object",
        "properties": {
          "recommendations": {
            "type": "number"
          },
          "reactions": {
            "type": "number"
          },
          "savedCards": {
            "type": "number"
          }
        },
        "required": [
          "recommendations",
          "reactions",
          "savedCards"
        ]
      },
      "AdminUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "lastActiveAt": {
            "type": "string",
            "nullable": true
          },
          "_count": {
            "$ref": "#/components/schemas/AdminUserCountsDto"
          }
        },
        "required": [
          "id",
          "email",
          "nickname",
          "role",
          "createdAt",
          "_count"
        ]
      },
      "AdminUsersPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUserResponseDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          }
        },
        "required": [
          "items"
        ]
      },
...
      "AdminRoomResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "topicTags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "visibility": {
            "enum": [
              "public",
              "private",
              "invite"
            ],
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "closed",
              "archived"
            ],
            "type": "string"
          },
          "memberCount": {
            "type": "number"
          },
          "passwordHint": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "owner": {
            "$ref": "#/components/schemas/AdminRoomOwnerDto"
          }
        },
        "required": [
          "id",
          "name",
          "topicTags",
          "visibility",
          "status",
          "memberCount",
          "createdAt",
          "updatedAt",
          "ownerId",
          "owner"
        ]
      },
      "AdminRoomsPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminRoomResponseDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          }
        },
        "required": [
          "items"
        ]
      },
      "UpdateAdminRoomDto": {
        "type": "object",
        "properties": {
          "status": {
            "enum": [
              "active",
              "closed",
              "archived"
            ],
            "type": "string"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "description": "방장 통지용 사유 · 닫기·보관 시 필수"
          }
        }
      },
...
      "DmLastMessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "body": {
            "type": "string"
          },
          "senderId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "body",
          "senderId",
          "createdAt"
        ]
      },
      "DmListItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "enum": [
              "pending",
              "open",
              "declined"
            ],
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "other": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthorSnippetDto"
              }
            ]
          },
          "lastMessage": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DmLastMessageDto"
              }
            ]
          },
          "unread": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "status",
          "updatedAt",
          "unread"
        ]
      },
      "DmRequestItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "enum": [
              "pending",
              "open",
              "declined"
            ],
            "type": "string"
          },
          "requestedById": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "createdAt": {
            "type": "string"
          },
          "other": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthorSnippetDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "status",
          "createdAt"
        ]
      },
      "DmDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "enum": [
              "pending",
              "open",
              "declined"
            ],
            "type": "string"
          },
          "requestedById": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "other": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthorSnippetDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "OpenDmDto": {
        "type": "object",
        "properties": {
          "otherUserId": {
            "type": "string",
            "format": "uuid",
            "description": "상대 userId"
          }
        },
        "required": [
          "otherUserId"
        ]
      },
      "DmMessageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "dmId": {
            "type": "string",
            "format": "uuid"
          },
          "senderId": {
            "type": "string",
            "format": "uuid"
          },
          "body": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "sender": {
            "$ref": "#/components/schemas/AuthorSnippetDto"
          }
        },
        "required": [
          "id",
          "dmId",
          "senderId",
          "body",
          "createdAt",
          "updatedAt",
          "sender"
        ]
      },
      "SendDmMessageDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "description": "메시지 내용"
          }
        },
        "required": [
          "body"
        ]
      },
...
      "RoomResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "topicTags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "visibility": {
            "enum": [
              "public",
              "private",
              "invite"
            ],
            "type": "string"
          },
          "passwordHint": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "enum": [
              "active",
              "closed",
              "archived"
            ],
            "type": "string"
          },
          "memberCount": {
            "type": "number"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/AuthorSnippetDto"
          },
          "lastMessageAt": {
            "type": "string",
            "nullable": true,
            "description": "`/rooms/mine`만"
          },
          "lastReadAt": {
            "type": "string"
          },
          "unread": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "topicTags",
          "visibility",
          "status",
          "memberCount",
          "ownerId",
          "createdAt",
          "updatedAt",
          "owner"
        ]
      },
...
      "UpdateRoomDto": {
        "type": "object",
        "properties": {
          "topicTags": {
            "maxItems": 8,
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 20
            }
          },
          "name": {
            "type": "string",
            "maxLength": 40
          },
          "description": {
            "type": "string",
            "nullable": true,
            "maxLength": 200
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "private",
              "invite"
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 64
          },
          "passwordHint": {
            "type": "string",
            "nullable": true,
            "maxLength": 40
          }
        }
      },
...
    }
  }
}